requirements.txt
#AI Orchestration
langgraph>=0.2.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.0

#web api
fastapi>=0.111.0
uvicorn[standard]>=0.34.0

#Databases
neo4j>=5.19.0
langchain-neo4j
qdrant-client>=1.9.0

#Data Ingestion
unstructured[all-docs]>=0.14.0
unstructured-client>=0.25.0
unstructured[pdf]
unstructured_inference>=0.7.0
earthengine-api
splink
datasets>=2.19.0
requests>=2.31.0

#Entity Resolution
rapidfuzz>=3.9.0

#UI and Tools
streamlit
python-dotenv>=1.0.0
pandas>=2.2.0
geopandas>=0.14.0
rasterio>=1.3.0

